### Overview

Add `class llama_kv_cache_unified_iswa` for interleaved SWA attention support.

The implementation internally utilizes 2 instances of the existing `llama_kv_cache_unified` - one for the non-SWA and one for the SWA layers of the model. To achieve that, the `llama_kv_cache_unified` implementation is updated to be able to cache a subset of the model's layers (instead of always caching all layers as it is on `master`). The 2 internal caches behave almost in exactly the same way with 2 main differences:

- The SWA cache is much smaller
- The SWA cache automatically "forgets/prunes" old tokens upon successful commit (i.e. successful batch decode)

The size of the SWA cache is computed as:

```
PAD(n_swa*n_seq_max + n_batch)
```

This way we can store the cache data for the last `n_swa` tokens for all sequences and we also have room to evaluate a new batch of tokens with size up to `n_batch`.

Note that advanced cache operations such as removing tokens or shifting their positions are not possible when using SWA cache, because token information becomes lost when the window slides. For such cases, we can "fallback" to the old implementation by expanding the SWA cache size to the full context and disabling the SWA token pruning. This of course would lead to more memory usage. See the `swa_full` flag for more info.


The new `llama_kv_cache_unified_iswa` can be used for non-SWA models with `n_swa = n_ctx_train`.

---

## Main changes

- Move KV cache store and view logic from `llama-graph` to `llama-kv-cache`
- Move KV cache mask creation logic from `llama-graph` to `llama-kv-cache`
- The inputs to `build_attn_mha()` are now not permuted
- The QKV self-attention code is now more harmonious:
  ```c++
    const llama_kv_cache_unified * kv_self = static_cast<const llama_kv_cache_unified *>(memory);

    // store to KV cache
    {
        ggml_build_forward_expand(gf, kv_self->cpy_k(ctx0, k_cur, il));
        ggml_build_forward_expand(gf, kv_self->cpy_v(ctx0, v_cur, il));
    }

    const auto & kq_mask = inp->get_kq_mask();

    ggml_tensor * q = q_cur;
    ggml_tensor * k = kv_self->get_k(ctx0, il);
    ggml_tensor * v = kv_self->get_v(ctx0, il);

    ggml_tensor * cur = build_attn_mha(gf, q, k, v, kq_b, kq_mask, v_mla, kq_scale);
    cb(cur, "kqv_out", il);
  ```
- Add `enum hparams.swa_type` to support chunked and non-chunked SWA (remove `hparams.n_attn_chunk`)
- Add `class llama_kv_cache_unified_iswa` - new iSWA cache that internally utilizes 2 standard `llama_kv_cache_unified` instances
- Make the `llama_kv_cache_unified` implementation more private and polish the interface
- Move the Llama 4 build function to a new `llm_build_llama_iswa()`
- `llama-server` now respects `llama_kv_self_can_shift(ctx)`
- The `llama_decode` now attempts to do a defrag if it fails to fit the input batch in the cache
- The `llama_decode` now correctly restores the cache state in all cases
- Examples can fallback to full-size SWA cache with `--swa-full`

## API changes

- Update `llama_context_params` - add `bool swa_full`

### TODO

- [x] Cut-off old SWA tokens in `llama_kv_cache_unified_iswa::commit()`
- [x] Pass `n_seq_max` and `n_batch` to the KV cache and utilize it to determine SWA cache size
- [x] Allow KV shift when SWA window size is big enough
- [x] Add limits to batch size based on SWA window
- [x] `llama-server` check for `llama_kv_self_can_shift`
- [x] Add context parameter for switching between small and large SWA cache (https://github.com/ggml-org/llama.cpp/pull/13194#issuecomment-2868343055)

### Testing

Any help with testing the following scenarios and reporting the results are highly appreciated:

- [x] Llama 4
- [x] Phi 3
- [x] Gemma 2
- [x] Gemma 3
- [x] Cohere 2
- [x] Multi-user
- [x] Context shift
- [x] Context reuse
- [x] Speculative decoding?

### Next PRs

- [ ] Split KV cache implementations in separate source files
- [ ] Remove `llama_kv_cache_view` API (not useful, can be replaced with internal debugging functions)
- [ ] Add `struct kv_cells` and simplify logic with modifying the cells
- [ ] Refactor the `llama_kv_cache` logic to allow SWA cache with size `n_swa + n_ubatch`
- [ ] Set defrag threshold to 0.0 by default
- [ ] `llama_decode` distinguish return code when we are sure that even after defrag there is no space available
- [ ] Update experimental status of `llama_context_params`
- [ ] Avoid `llama_kv_cache::set_full()`
- [ ] Rework `llama_kv_cache` to not maintain the batching state (https://github.com/ggml-org/llama.cpp/pull/13194#pullrequestreview-2852087639)
- [ ] Consider `template <bool SWA> llm_build_llama()`

---

<details>
<summary>outdated</summary>

This is still very WIP - the goal is to redesign the unified KV cache to properly support layers with sliding-window attention (SWA) in order to reduce the memory usage for models such as Gemma3.

However, while working on this, I realized that enabling this option would prevent context caching, which IMO is a pretty big deal. So I am wondering if I am missing something.

The reason we cannot do context caching with SWA enabled is because when the window slides, we "forget" the old KV stuff and there is no way to recover it without recomputing it. This means, no prefix cache in `llama-server` (ok, just last-prefix caching works), no context shift, no context reuse, etc. So I am having some doubts if this is really worth supporting.

Any thoughts?
</details>