@ggerganov 2025-04-29
> It's not very clear to me how to handle SWA with a unified cache where there may be multiple sequences, and it is not always obvious what tokens can be dropped from the cache. However I think it is definitely worth it for the single user case, which after all is the main use case of llama.cpp.

The way I am approaching it is to have the "KV cells" information maintained separately for the non-SWA and SWA layers. This way, upon each KV cache commit (see #12799), we can do a pass over the SWA cells and automatically remove those that have position `pos < pos_max(seq_id) - n_swa`. Note that such tokens are only pruned from the SWA cells, while they remain in the non-SWA cells. When constructing the KQ mask for the graph, we use the non-SWA cells to construct the `kq_mask` and the SWA cells to construct the `kq_mask_swa`. 

The rest of the logic is the same - it just operates on both set of cells. For example, `find_slot` searches in both the non-SWA and SWA cells.