@JohannesGaessler 2025-04-29
My experience with the Gemma models in the context of [Elo HeLLM](https://github.com/JohannesGaessler/elo_hellm) has been that they required a disproportionate amount of computational resources to run benchmarks. The reason is that I was able to fit comparatively fewer parallel slots on 1 or 2 GPUs and my throughput was lower as a consequence. At least for my use case I value low memory usage for the context more than I value prompt caching because I have O(10000) short prompts and I'm bottlenecked mostly by generation throughput.