@ggerganov 2025-04-30
Continuing thinking about the logic for when to discard tokens from the cache, it's indeed tricky and not very clear how to do. For example, when doing speculative decoding, we can submit a draft batch with `D` tokens to the target model. If we apply the pruning logic from my [previous comment](https://github.com/ggml-org/llama.cpp/pull/13194#issuecomment-2839971403) strictly, then this would cause to "forget" `D-1` of the oldest tokens in the SWA layers, which depending if the draft gets rejected would be problematic. This makes me think that we should probably have some "extra room" in the SWA cache - for example `n_swa + 2*n_batch`. And the prune logic should be something like: `pos < pos_max(seq_id) - n_swa - n_batch`.