@ngxson 2025-04-30
> This makes me think that we should probably have some "extra room" in the SWA cache - for example `n_swa + 2*n_batch`. And the prune logic should be something like: `pos < pos_max(seq_id) - n_swa - n_batch`.

Yes I was thinking about this too, I think it can be a bit complicated to manage this case, but totally possible.

We can let user specify how many tokens are allocated in the sliding layers. For example, given `n_swa=512`, if llama_context is created with `n_ctx=4096` and `n_ctx_swa=1024`, this will allow user to rollback until `n_past - (1024 - 512)`

We can further let `n_ctx_swa = n_ctx * scale` by default to make it transparent to end-user, with `scale=0.5` by default for example. If `scale=-1` then `n_ctx_swa=n_swa`

And finally, we may need to add an API to return the furthest `n_past` that user can rollback to, maybe something like `llama_kv_self_get_minimum_pos` ?