@isaac-mcfadyen 2025-04-30
I'd +1 the ability to allow the user to switch.

Some use-cases benefit greatly from the prefix caching (example: on Metal systems with 48GB of RAM/VRAM, where pp is much slower than tg and we have plenty of VRAM anyway) so allowing the user to choose would be optimal.