@LostRuins 2025-05-10
Some people recently mentioned concerns with this PR  - I think caching is quite important for a subset of users who don't have GPUs and run purely CPU only. 

They are fine spending initial minutes or more ingesting a large initial prompts which they then reuse for many future turns -  generation speed itself is usable, but the inability to cache would be crippling for such users.