@ggerganov 2025-05-10
Both the old cache (i.e. more memory usage, but with advanced caching supported) and the new cache (less memory with just last-prefix caching) will be supported. Still figuring the implementation details - will likely be supported via a flag or a [parameter](https://github.com/ggml-org/llama.cpp/pull/13194#issuecomment-2841386823).