@windswand 2026-02-18
Running Unsloth's Qwen3.5-397B-A17B Q6_K_XL GGUF on latest llama.cpp (with [pwilkin/autoparser](https://github.com/pwilkin/llama.cpp.git) tool calling fixes) and it throws errors directing me to this issue. I'm running Claude cli via LiteLLM into llama.cpp.

This is the error: `forcing full prompt re-processing due to lack of cache data (likely due to SWA or hybrid/recurrent memory, see https://github.com/ggml-org/llama.cpp/pull/13194#issuecomment-2868343055)`

Full log:

```
> build/bin/llama-server --host 0.0.0.0 --port 8080 -m ~/.cache/huggingface/hub/models--unsloth--Qwen3.5-397B-A17B-GGUF/snapshots/59c9e42d5a7a6edc5cc7fb35cb8237b4a7cc86cd/UD-Q6_K_XL/Qwen3.5-397B-A17B-UD-Q6_K_XL-00001-of-00008.gguf -ngl 999 --threads 127 -fa on -c 196608 --temp 0.7 --top-p 0.8 --top-k 20 --min-p 0.00 --chat-template-kwargs "{\"enable_thinking\": false}"
ggml_cuda_init: found 4 CUDA devices:
  Device 0: NVIDIA RTX PRO 6000 Blackwell Workstation Edition, compute capability 12.0, VMM: yes
  Device 1: NVIDIA RTX PRO 6000 Blackwell Workstation Edition, compute capability 12.0, VMM: yes
  Device 2: NVIDIA RTX PRO 6000 Blackwell Workstation Edition, compute capability 12.0, VMM: yes
  Device 3: NVIDIA RTX PRO 6000 Blackwell Workstation Edition, compute capability 12.0, VMM: yes
main: n_parallel is set to auto, using n_parallel = 4 and kv_unified = true
build: 8101 (519e1f1f3) with GNU 13.3.0 for Linux x86_64
system info: n_threads = 127, n_threads_batch = 127, total_threads = 256

system_info: n_threads = 127 (n_threads_batch = 127) / 256 | CUDA : ARCHS = 1200 | USE_GRAPHS = 1 | PEER_MAX_BATCH_SIZE = 128 | FA_ALL_QUANTS = 1 | BLACKWELL_NATIVE_FP4 = 1 | CPU : SSE3 = 1 | SSSE3 = 1 | AVX = 1 | AVX_VNNI = 1 | AVX2 = 1 | F16C = 1 | FMA = 1 | BMI2 = 1 | AVX512 = 1 | AVX512_VBMI = 1 | AVX512_VNNI = 1 | AVX512_BF16 = 1 | LLAMAFILE = 1 | OPENMP = 1 | REPACK = 1 |

Running without SSL
init: using 255 threads for HTTP server
start: binding port with default address family
main: loading model
srv    load_model: loading model '/home/womble/.cache/huggingface/hub/models--unsloth--Qwen3.5-397B-A17B-GGUF/snapshots/59c9e42d5a7a6edc5cc7fb35cb8237b4a7cc86cd/UD-Q6_K_XL/Qwen3.5-397B-A17B-UD-Q6_K_XL-00001-of-00008.gguf'
common_init_result: fitting params to device memory, for bugs during this step try to reproduce them with -fit off, or provide --verbose logs if the bug only occurs with -fit on
llama_params_fit_impl: projected memory use with initial parameters [MiB]:
llama_params_fit_impl:   - CUDA0 (NVIDIA RTX PRO 6000 Blackwell Workstation Edition):  97249 total,  87086 used,   9403 free vs. target of   1024
llama_params_fit_impl:   - CUDA1 (NVIDIA RTX PRO 6000 Blackwell Workstation Edition):  97249 total,  82788 used,  13701 free vs. target of   1024
llama_params_fit_impl:   - CUDA2 (NVIDIA RTX PRO 6000 Blackwell Workstation Edition):  97249 total,  79982 used,  16523 free vs. target of   1024
llama_params_fit_impl:   - CUDA3 (NVIDIA RTX PRO 6000 Blackwell Workstation Edition):  97249 total,  80341 used,  16181 free vs. target of   1024
llama_params_fit_impl: projected to use 330198 MiB of device memory vs. 386009 MiB of free device memory
llama_params_fit_impl: targets for free memory can be met on all devices, no changes needed
llama_params_fit: successfully fit params to free device memory
llama_params_fit: fitting params to free memory took 0.70 seconds
llama_model_load_from_file_impl: using device CUDA0 (NVIDIA RTX PRO 6000 Blackwell Workstation Edition) (0000:01:00.0) - 96689 MiB free
llama_model_load_from_file_impl: using device CUDA1 (NVIDIA RTX PRO 6000 Blackwell Workstation Edition) (0000:21:00.0) - 96689 MiB free
llama_model_load_from_file_impl: using device CUDA2 (NVIDIA RTX PRO 6000 Blackwell Workstation Edition) (0000:41:00.0) - 96689 MiB free
llama_model_load_from_file_impl: using device CUDA3 (NVIDIA RTX PRO 6000 Blackwell Workstation Edition) (0000:c1:00.0) - 96689 MiB free
llama_model_loader: additional 7 GGUFs metadata loaded.
llama_model_loader: loaded meta data with 48 key-value pairs and 1098 tensors from /home/womble/.cache/huggingface/hub/models--unsloth--Qwen3.5-397B-A17B-GGUF/snapshots/59c9e42d5a7a6edc5cc7fb35cb8237b4a7cc86cd/UD-Q6_K_XL/Qwen3.5-397B-A17B-UD-Q6_K_XL-00001-of-00008.gguf (version GGUF V3 (latest))
llama_model_loader: Dumping metadata keys/values. Note: KV overrides do not apply in this output.
llama_model_loader: - kv   0:                       general.architecture str              = qwen35moe
llama_model_loader: - kv   1:                               general.type str              = model
llama_model_loader: - kv   2:                               general.name str              = Qwen3.5-397B-A17B
llama_model_loader: - kv   3:                           general.basename str              = Qwen3.5-397B-A17B
llama_model_loader: - kv   4:                       general.quantized_by str              = Unsloth
llama_model_loader: - kv   5:                         general.size_label str              = 397B-A17B
llama_model_loader: - kv   6:                            general.license str              = apache-2.0
llama_model_loader: - kv   7:                       general.license.link str              = https://huggingface.co/Qwen/Qwen3.5-3...
llama_model_loader: - kv   8:                           general.repo_url str              = https://huggingface.co/unsloth
llama_model_loader: - kv   9:                   general.base_model.count u32              = 1
llama_model_loader: - kv  10:                  general.base_model.0.name str              = Qwen3.5 397B A17B
llama_model_loader: - kv  11:          general.base_model.0.organization str              = Qwen
llama_model_loader: - kv  12:              general.base_model.0.repo_url str              = https://huggingface.co/Qwen/Qwen3.5-3...
llama_model_loader: - kv  13:                               general.tags arr[str,2]       = ["unsloth", "image-text-to-text"]
llama_model_loader: - kv  14:                      qwen35moe.block_count u32              = 60
llama_model_loader: - kv  15:                   qwen35moe.context_length u32              = 262144
llama_model_loader: - kv  16:                 qwen35moe.embedding_length u32              = 4096
llama_model_loader: - kv  17:             qwen35moe.attention.head_count u32              = 32
llama_model_loader: - kv  18:          qwen35moe.attention.head_count_kv u32              = 2
llama_model_loader: - kv  19:          qwen35moe.rope.dimension_sections arr[i32,4]       = [11, 11, 10, 0]
llama_model_loader: - kv  20:                   qwen35moe.rope.freq_base f32              = 10000000.000000
llama_model_loader: - kv  21: qwen35moe.attention.layer_norm_rms_epsilon f32              = 0.000001
llama_model_loader: - kv  22:                     qwen35moe.expert_count u32              = 512
llama_model_loader: - kv  23:                qwen35moe.expert_used_count u32              = 10
llama_model_loader: - kv  24:             qwen35moe.attention.key_length u32              = 256
llama_model_loader: - kv  25:           qwen35moe.attention.value_length u32              = 256
llama_model_loader: - kv  26:       qwen35moe.expert_feed_forward_length u32              = 1024
llama_model_loader: - kv  27: qwen35moe.expert_shared_feed_forward_length u32              = 1024
llama_model_loader: - kv  28:                  qwen35moe.ssm.conv_kernel u32              = 4
llama_model_loader: - kv  29:                   qwen35moe.ssm.state_size u32              = 128
llama_model_loader: - kv  30:                  qwen35moe.ssm.group_count u32              = 16
llama_model_loader: - kv  31:               qwen35moe.ssm.time_step_rank u32              = 64
llama_model_loader: - kv  32:                   qwen35moe.ssm.inner_size u32              = 8192
llama_model_loader: - kv  33:          qwen35moe.full_attention_interval u32              = 4
llama_model_loader: - kv  34:             qwen35moe.rope.dimension_count u32              = 64
llama_model_loader: - kv  35:                       tokenizer.ggml.model str              = gpt2
llama_model_loader: - kv  36:                         tokenizer.ggml.pre str              = qwen35
llama_model_loader: - kv  37:                      tokenizer.ggml.tokens arr[str,248320]  = ["!", "\"", "#", "$", "%", "&", "'", ...
llama_model_loader: - kv  38:                  tokenizer.ggml.token_type arr[i32,248320]  = [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ...
llama_model_loader: - kv  39:                      tokenizer.ggml.merges arr[str,247587]  = ["Ġ Ġ", "ĠĠ ĠĠ", "i n", "Ġ t",...
llama_model_loader: - kv  40:                tokenizer.ggml.eos_token_id u32              = 248046
llama_model_loader: - kv  41:            tokenizer.ggml.padding_token_id u32              = 248055
llama_model_loader: - kv  42:                    tokenizer.chat_template str              = {%- set image_count = namespace(value...
llama_model_loader: - kv  43:               general.quantization_version u32              = 2
llama_model_loader: - kv  44:                          general.file_type u32              = 18
llama_model_loader: - kv  45:                                   split.no u16              = 0
llama_model_loader: - kv  46:                        split.tensors.count i32              = 1098
llama_model_loader: - kv  47:                                split.count u16              = 8
llama_model_loader: - type  f32:  451 tensors
llama_model_loader: - type q8_0:  105 tensors
llama_model_loader: - type q6_K:  542 tensors
print_info: file format = GGUF V3 (latest)
print_info: file type   = Q6_K
print_info: file size   = 305.62 GiB (6.62 BPW)
load: 0 unused tokens
load: printing all EOG tokens:
load:   - 248044 ('<|endoftext|>')
load:   - 248046 ('<|im_end|>')
load:   - 248063 ('<|fim_pad|>')
load:   - 248064 ('<|repo_name|>')
load:   - 248065 ('<|file_sep|>')
load: special tokens cache size = 33
load: token to piece cache size = 1.7581 MB
print_info: arch                  = qwen35moe
print_info: vocab_only            = 0
print_info: no_alloc              = 0
print_info: n_ctx_train           = 262144
print_info: n_embd                = 4096
print_info: n_embd_inp            = 4096
print_info: n_layer               = 60
print_info: n_head                = 32
print_info: n_head_kv             = 2
print_info: n_rot                 = 64
print_info: n_swa                 = 0
print_info: is_swa_any            = 0
print_info: n_embd_head_k         = 256
print_info: n_embd_head_v         = 256
print_info: n_gqa                 = 16
print_info: n_embd_k_gqa          = 512
print_info: n_embd_v_gqa          = 512
print_info: f_norm_eps            = 0.0e+00
print_info: f_norm_rms_eps        = 1.0e-06
print_info: f_clamp_kqv           = 0.0e+00
print_info: f_max_alibi_bias      = 0.0e+00
print_info: f_logit_scale         = 0.0e+00
print_info: f_attn_scale          = 0.0e+00
print_info: n_ff                  = 0
print_info: n_expert              = 512
print_info: n_expert_used         = 10
print_info: n_expert_groups       = 0
print_info: n_group_used          = 0
print_info: causal attn           = 1
print_info: pooling type          = 0
print_info: rope type             = 40
print_info: rope scaling          = linear
print_info: freq_base_train       = 10000000.0
print_info: freq_scale_train      = 1
print_info: n_ctx_orig_yarn       = 262144
print_info: rope_yarn_log_mul     = 0.0000
print_info: rope_finetuned        = unknown
print_info: mrope sections        = [11, 11, 10, 0]
print_info: ssm_d_conv            = 4
print_info: ssm_d_inner           = 8192
print_info: ssm_d_state           = 128
print_info: ssm_dt_rank           = 64
print_info: ssm_n_group           = 16
print_info: ssm_dt_b_c_rms        = 0
print_info: model type            = ?B
print_info: model params          = 396.35 B
print_info: general.name          = Qwen3.5-397B-A17B
print_info: vocab type            = BPE
print_info: n_vocab               = 248320
print_info: n_merges              = 247587
print_info: BOS token             = 11 ','
print_info: EOS token             = 248046 '<|im_end|>'
print_info: EOT token             = 248046 '<|im_end|>'
print_info: PAD token             = 248055 '<|vision_pad|>'
print_info: LF token              = 198 'Ċ'
print_info: FIM PRE token         = 248060 '<|fim_prefix|>'
print_info: FIM SUF token         = 248062 '<|fim_suffix|>'
print_info: FIM MID token         = 248061 '<|fim_middle|>'
print_info: FIM PAD token         = 248063 '<|fim_pad|>'
print_info: FIM REP token         = 248064 '<|repo_name|>'
print_info: FIM SEP token         = 248065 '<|file_sep|>'
print_info: EOG token             = 248044 '<|endoftext|>'
print_info: EOG token             = 248046 '<|im_end|>'
print_info: EOG token             = 248063 '<|fim_pad|>'
print_info: EOG token             = 248064 '<|repo_name|>'
print_info: EOG token             = 248065 '<|file_sep|>'
print_info: max token length      = 256
load_tensors: loading model tensors, this can take a while... (mmap = true, direct_io = false)
load_tensors: offloading output layer to GPU
load_tensors: offloading 59 repeating layers to GPU
load_tensors: offloaded 61/61 layers to GPU
load_tensors:   CPU_Mapped model buffer size =  1030.62 MiB
load_tensors:        CUDA0 model buffer size = 83382.71 MiB
load_tensors:        CUDA1 model buffer size = 77225.60 MiB
load_tensors:        CUDA2 model buffer size = 77215.03 MiB
load_tensors:        CUDA3 model buffer size = 74104.63 MiB
....................................................................................................
common_init_result: added <|endoftext|> logit bias = -inf
common_init_result: added <|im_end|> logit bias = -inf
common_init_result: added <|fim_pad|> logit bias = -inf
common_init_result: added <|repo_name|> logit bias = -inf
common_init_result: added <|file_sep|> logit bias = -inf
llama_context: constructing llama_context
llama_context: n_seq_max     = 4
llama_context: n_ctx         = 196608
llama_context: n_ctx_seq     = 196608
llama_context: n_batch       = 2048
llama_context: n_ubatch      = 512
llama_context: causal_attn   = 1
llama_context: flash_attn    = enabled
llama_context: kv_unified    = true
llama_context: freq_base     = 10000000.0
llama_context: freq_scale    = 1
llama_context: n_ctx_seq (196608) < n_ctx_train (262144) -- the full capacity of the model will not be utilized
llama_context:  CUDA_Host  output buffer size =     3.79 MiB
llama_kv_cache:      CUDA0 KV buffer size =  1536.00 MiB
llama_kv_cache:      CUDA1 KV buffer size =  1152.00 MiB
llama_kv_cache:      CUDA2 KV buffer size =  1536.00 MiB
llama_kv_cache:      CUDA3 KV buffer size =  1536.00 MiB
llama_kv_cache: size = 5760.00 MiB (196608 cells,  15 layers,  4/1 seqs), K (f16): 2880.00 MiB, V (f16): 2880.00 MiB
llama_memory_recurrent:      CUDA0 RS buffer size =   198.75 MiB
llama_memory_recurrent:      CUDA1 RS buffer size =   198.75 MiB
llama_memory_recurrent:      CUDA2 RS buffer size =   182.19 MiB
llama_memory_recurrent:      CUDA3 RS buffer size =   165.62 MiB
llama_memory_recurrent: size =  745.31 MiB (     4 cells,  60 layers,  4 seqs), R (f32):   25.31 MiB, S (f32):  720.00 MiB
llama_context: pipeline parallelism enabled
sched_reserve: reserving ...
sched_reserve:      CUDA0 compute buffer size =  1968.67 MiB
sched_reserve:      CUDA1 compute buffer size =  4179.75 MiB
sched_reserve:      CUDA2 compute buffer size =  1017.38 MiB
sched_reserve:      CUDA3 compute buffer size =  4503.63 MiB
sched_reserve:  CUDA_Host compute buffer size =  1552.08 MiB
sched_reserve: graph nodes  = 11279 (with bs=512), 7094 (with bs=1)
sched_reserve: graph splits = 9
sched_reserve: reserve took 459.56 ms, sched copies = 4
common_init_from_params: warming up the model with an empty run - please wait ... (--no-warmup to disable)
srv    load_model: initializing slots, n_slots = 4
common_speculative_is_compat: the target context does not support partial sequence removal
srv    load_model: speculative decoding not supported by this context
slot   load_model: id  0 | task -1 | new slot, n_ctx = 196608
slot   load_model: id  1 | task -1 | new slot, n_ctx = 196608
slot   load_model: id  2 | task -1 | new slot, n_ctx = 196608
slot   load_model: id  3 | task -1 | new slot, n_ctx = 196608
srv    load_model: prompt cache is enabled, size limit: 8192 MiB
srv    load_model: use `--cache-ram 0` to disable the prompt cache
srv    load_model: for more info see https://github.com/ggml-org/llama.cpp/pull/16391
init: chat template, example_format: '<|im_start|>system
You are a helpful assistant<|im_end|>
<|im_start|>user
Hello<|im_end|>
<|im_start|>assistant
Hi there<|im_end|>
<|im_start|>user
How are you?<|im_end|>
<|im_start|>assistant
<think>

</think>

'
srv          init: init: chat template, thinking = 1
main: model loaded
main: server is listening on http://0.0.0.0:8080
main: starting the main loop...
srv  update_slots: all slots are idle
srv  params_from_: Chat format: peg-native
slot get_availabl: id  3 | task -1 | selected slot by LRU, t_last = -1
slot launch_slot_: id  3 | task -1 | sampler chain: logits -> ?penalties -> ?dry -> ?top-n-sigma -> top-k -> ?typical -> top-p -> ?min-p -> ?xtc -> temp-ext -> dist
slot launch_slot_: id  3 | task 0 | processing task, is_child = 0
slot update_slots: id  3 | task 0 | new prompt, n_ctx_slot = 196608, n_keep = 0, task.n_tokens = 428
slot update_slots: id  3 | task 0 | n_tokens = 0, memory_seq_rm [0, end)
slot update_slots: id  3 | task 0 | prompt processing progress, n_tokens = 428, batch.n_tokens = 428, progress = 1.000000
slot update_slots: id  3 | task 0 | prompt done, n_tokens = 428, batch.n_tokens = 428
slot init_sampler: id  3 | task 0 | init sampler, took 0.12 ms, tokens: text = 428, total = 428
slot print_timing: id  3 | task 0 |
prompt eval time =     595.81 ms /   428 tokens (    1.39 ms per token,   718.35 tokens per second)
       eval time =     242.43 ms /    15 tokens (   16.16 ms per token,    61.87 tokens per second)
      total time =     838.23 ms /   443 tokens
slot      release: id  3 | task 0 | stop processing: n_tokens = 442, truncated = 0
srv  update_slots: all slots are idle
srv  log_server_r: done request: POST /v1/messages 127.0.0.1 200
srv  params_from_: Chat format: peg-native
slot get_availabl: id  3 | task -1 | selected slot by LCP similarity, sim_best = 0.169 (> 0.100 thold), f_keep = 0.048
srv  get_availabl: updating prompt cache
srv   prompt_save:  - saving prompt with length 442, total state size = 199.284 MiB
srv  params_from_: Chat format: peg-native
srv  params_from_: Chat format: peg-native
srv          load:  - looking for better prompt, base f_keep = 0.048, sim = 0.169
srv        update:  - cache state: 1 prompts, 199.284 MiB (limits: 8192.000 MiB, 196608 tokens, 196608 est)
srv        update:    - prompt 0x5a84ff17a710:     442 tokens, checkpoints:  0,   199.284 MiB
srv  get_availabl: prompt cache update took 71.59 ms
slot launch_slot_: id  3 | task -1 | sampler chain: logits -> ?penalties -> ?dry -> ?top-n-sigma -> top-k -> ?typical -> top-p -> ?min-p -> ?xtc -> temp-ext -> dist
slot launch_slot_: id  3 | task 16 | processing task, is_child = 0
slot get_availabl: id  2 | task -1 | selected slot by LRU, t_last = -1
slot launch_slot_: id  2 | task -1 | sampler chain: logits -> ?penalties -> ?dry -> ?top-n-sigma -> top-k -> ?typical -> top-p -> ?min-p -> ?xtc -> temp-ext -> dist
slot launch_slot_: id  2 | task 17 | processing task, is_child = 0
slot get_availabl: id  1 | task -1 | selected slot by LRU, t_last = -1
slot launch_slot_: id  1 | task -1 | sampler chain: logits -> ?penalties -> ?dry -> ?top-n-sigma -> top-k -> ?typical -> top-p -> ?min-p -> ?xtc -> temp-ext -> dist
…